+++ spec/dungeon_spec.rb
@@ -1,20 +1,11 @@
-def convert(letter_in)
-  letters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-  numberFromLetter = letters.index(letter_in)
-  if numberFromLetter == nil
-    numberOfLetter = letter_in.to_i
-    return numberOfLetter + 5 if numberOfLetter < 4
-    return numberOfLetter
-  end
-  return numberFromLetter - 25 if numberFromLetter > 24
-  return numberFromLetter - 20 if numberFromLetter > 19
-  return numberFromLetter - 15 if numberFromLetter > 14
-  return numberFromLetter - 10 if numberFromLetter > 9
-  return numberFromLetter - 5 if numberFromLetter > 4
-  numberFromLetter
-end
+require '../lib/dungeon.rb'
 
 describe "The Dungeon" do
+
+  before do
+    @dungeon = Dungeon.new
+  end
+
   xit "gives an object based on a name" do
     expect(objectTo("Gideon")).to be("Orb")
   end
@@ -33,7 +24,7 @@ describe "The Dungeon" do
       "Z" => 0
     }.each_pair do |letter, number|
       it "converts #{letter} to #{number}" do
-        expect(convert(letter)).to be(number)
+        expect(@dungeon.convert(letter)).to be(number)
       end
     end
   end
@@ -46,7 +37,7 @@ describe "The Dungeon" do
       "6" => 6
     }.each_pair do |number, toNumber|
       it "converts #{number} to #{toNumber}" do
-        expect(convert(number)).to be(toNumber)
+        expect(@dungeon.convert(number)).to be(toNumber)
       end
     end
   end
